move isScoreable check

+++ app/Models/Beatmapset.php
@@ -20,6 +20,7 @@
 
 namespace App\Models;
 
+use App\Exceptions\AuthorizationException;
 use App\Exceptions\BeatmapProcessorException;
 use App\Jobs\CheckBeatmapsetCovers;
 use App\Jobs\EsIndexDocument;
@@ -960,6 +961,10 @@ class Beatmapset extends Model implements AfterCommit
 
     public function delete()
     {
+        if ($this->isScoreable()) {
+            throw new AuthorizationException('This beatmap is no longer deleteable.');
+        }
+
         // kinda pointless since laravel doesn't error on delete
         return $this->getConnection()->transaction(function () {
             $this->beatmaps()->delete();
